Moved bundle_id to ids and aggregated ExtensionPointID onto enumeration class
