+++ force_bdss/base_extension_plugin.py
@@ -1,6 +1,7 @@
 from envisage.plugin import Plugin
 from traits.trait_types import List
 
+from .ids import ExtensionPointID
 from .data_sources.i_data_source_bundle import IDataSourceBundle
 from .kpi.i_kpi_calculator_bundle import IKPICalculatorBundle
 from .mco.i_multi_criteria_optimizer_bundle import \
@@ -25,17 +26,17 @@ class BaseExtensionPlugin(Plugin):
     #: A list of available Multi Criteria Optimizers this plugin exports.
     mco_bundles = List(
         IMultiCriteriaOptimizerBundle,
-        contributes_to='force.bdss.mco.bundles'
+        contributes_to=ExtensionPointID.MCO_BUNDLES
     )
 
     #: A list of the available Data Sources this plugin exports.
     data_source_bundles = List(
         IDataSourceBundle,
-        contributes_to='force.bdss.data_sources.bundles'
+        contributes_to=ExtensionPointID.DATA_SOURCE_BUNDLES
     )
 
     #: A list of the available KPI calculators this plugin exports.
     kpi_calculator_bundles = List(
         IKPICalculatorBundle,
-        contributes_to='force.bdss.kpi_calculators.bundles'
+        contributes_to=ExtensionPointID.KPI_CALCULATOR_BUNDLES
     )
